Moved list into new <Todo List> component. Changed ReactDOM.render to ReactDOM.createRoot()
